requirements.txt
openai
python-dotenv
pypdf
chromadb
sentence-transformers